Refactored image aspects as screen aspects
Necessary to keep all of this screen/image/display logic together. I originally used `Images` as the namespace because I thought this would end up being generic image handling but this has become more specialized for device screens (even though they are images). Keeping all of this together reduces the number of namespaces used at least.

Milestone: patch
