+++ lib/terminus/screens/toucher.rb
@@ -4,7 +4,7 @@ require "refinements/pathname"
 
 module Terminus
   # Touches the oldest file to make it the latest file.
-  module Images
+  module Screens
     using Refinements::Pathname
 
     Toucher = -> root { Pathname(root).files.min_by(&:ctime).touch }
